some number commit

+++ src/Components/NavBar.js
@@ -6,7 +6,9 @@ import { Redirect } from 'react-router-dom';
 export default class NavBar extends React.Component {
     // need to make this bigger
 
-    state = { path: null }
+    state = { 
+        path: null
+    }
 
     loggedIn = () => {
         // here is where we display Log In if no user ... or Log Out if user logged in
@@ -14,10 +16,17 @@ export default class NavBar extends React.Component {
     
     handleClick = (e) => {
         this.setState({ path: e.target.id })
+        if(e.target.id === 'logout'){
+            localStorage.clear()
+            // this.props.clearUser()
+            this.setState({ path: 'login'})
+        }
     }
 
-    handleLogOut = () => {
-        localStorage.clear(); 
+    
+
+    renderLogButton = () => {
+        return localStorage.length > 0 ? <Menu.Item name='logout' id='logout' onClick={this.handleClick} /> : <Menu.Item name='login' id='login' onClick={this.handleClick} />
     }
 
     render(){
@@ -30,9 +39,11 @@ export default class NavBar extends React.Component {
         return (
             <Sticky>
                 <Menu secondary inverted size="massive">
-                    <Menu.Item name='home' id='main' onClick={this.handleClick}/>
+                    <Menu.Item name='home' id='' onClick={this.handleClick}/>
                     <Menu.Item name='gear' id='gear' onClick={this.handleClick}/>
                     <Menu.Item name='resorts' id='resorts' onClick={this.handleClick}/>
+                    <Menu.Item name="post" id="post" onClick={this.handleClick}/>
+                    <Menu.Item name='users' id='users' onClick={this.handleClick} />
                    
            
                     <Menu.Menu position='right'>
@@ -40,7 +51,7 @@ export default class NavBar extends React.Component {
                         {/* i lose the semantic if i put the listener on the Icon ... but now I can only click around the Icon */}
                             <Icon name='user' />
                         </Menu.Item>
-                        <Menu.Item name='logout' id='logout' onClick={this.handleLogOut} />
+                        {this.renderLogButton()}
                     </Menu.Menu>
                 </Menu>
             </Sticky>
